클라이언트 수정
전투씬에서 다른 플레이어가 로딩중일때까지 씬을 렌더링 하지 않던 구조에서 클라이언트의 로딩이 끝나면 렌더링을 하고 다른 플레이어를 기다리는 구조로 변경

+++ Project Client/PillsFighter/Font.cpp
@@ -147,7 +147,7 @@ void CFont::Initialize(ID3D12Device *pd3dDevice, ID3D12GraphicsCommandList *pd3d
 	if (m_vTextureInfo.size() > 1) nType = RESOURCE_TEXTURE2D_ARRAY;
 	else nType = RESOURCE_TEXTURE2D;
 
-	m_pFontTexture = new CTexture(m_vTextureInfo.size(), nType, 0);
+	m_pFontTexture = new CTexture((int)m_vTextureInfo.size(), nType, 0);
 
 	for (int i = 0; i < m_vTextureInfo.size(); i++)
 	{
